authorization gate

+++ app/Http/Controllers/CommentController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 use App\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -16,6 +17,7 @@ class CommentController extends Controller
         $comment = new Comment;
         $comment->content=request()->content;
         $comment->article_id=request()->article_id;
+        $comment->user_id->auth()->user()->id;
         $comment->save();
 
         return back();
@@ -24,7 +26,13 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
+        if(Gate::allows('comment-delete',$comment)){
+            $comment->delete();
+            return back();
+        } else {
+            return back()->with('error','Unauthorize');
+        }
+        
 
         return back();
     }
